Also run tests on non-UTC time zones.
The tests are currently failing for zones west of UTC due to assumptions
related to `time.localtime`. The test suite should pass when run on
developer machines, which aren't usually set to UTC, so add explicit
cases to the testing matrix to ensure that once the issue is fixed, it
will not have regressions.

New_York and Sydney are chosen arbitrarily, but one represents a zone in
the southern hemisphere and one represents a zone west of UTC.