Fix tests on platforms with low clock resolution
On platforms without a high-resolution clock, such as Alpha and PA-RISC
is is likely that two sequential calls to time.time() will return the
same value if the execution time is not sufficient to allow one full
clock resolution cycle to pass.  This adds sleeps of one cycle to
enforce that the minimum amount of time to guarantee a clock change has
passed.

On systems with high-resolution clocks, clock_getres() will return 1ns;
in reality the sleep will take longer than 1ns to execute but should
still be a negligible amount of time.